RG-44 Delete own solutions not other's

+++ server/src/main/java/com/rolaface/controllers/CauseController.java
@@ -1,8 +1,10 @@
 package com.rolaface.controllers;
 
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,8 +17,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.rolaface.entities.Cause;
 import com.rolaface.entities.FlexErrorCause;
+import com.rolaface.entities.User;
+import com.rolaface.model.ContextUser;
 import com.rolaface.services.CauseService;
 import com.rolaface.services.FlexErrorCauseService;
+import com.rolaface.services.UserService;
 
 @CrossOrigin(origins = "http://localhost:4200", maxAge = 3600)
 @RestController
@@ -29,9 +34,19 @@ public class CauseController {
 	@Autowired
 	private FlexErrorCauseService errorCauseService;
 
+	@Autowired
+	private UserService userService;
+
 	@PostMapping
 	public Cause create(@RequestBody CauseDTO causeDTO) {
-		Cause cause = causeService.create(causeDTO.getCause());
+		int userId = ((ContextUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId();
+		User user = userService.findById(userId);
+
+		Cause cause = causeDTO.getCause();
+		cause.setUser(user);
+		cause.setCreatedTimestamp(new Date());
+		cause = causeService.create(cause);
+
 		FlexErrorCause errorCause = new FlexErrorCause();
 		errorCause.setCauseid(cause.getCauseid());
 		errorCause.setErrid(causeDTO.getErrid());
@@ -46,12 +61,21 @@ public class CauseController {
 
 	@PutMapping(path = { "/{id}" })
 	public Cause update(@RequestBody Cause cause) {
-		return causeService.update(cause);
+		int userId = ((ContextUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId();
+		if (userId == cause.getUser().getUserid()) {
+			cause = causeService.update(cause);
+		}
+		return cause;
 	}
 
 	@DeleteMapping(path = { "/{id}" })
 	public Cause delete(@PathVariable("id") int id) {
-		return causeService.delete(id);
+		int userId = ((ContextUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUserId();
+		Cause cause = causeService.findById(id);
+		if (userId == cause.getUser().getUserid()) {
+			cause = causeService.delete(id);
+		}
+		return cause;
 	}
 
 	@GetMapping
